set the right image for the release
At the moment in case of an release the image gets set to a default value of image: op-skr-registry.localhost:8888/unsigned/manager-images/nats-manager:<version> in the nats-manager.yaml due to an unset IMG var. This commit changes that by setting the IMG. Also, the script was made more generic by removing references to nats and instead, pass the string nats as a var from the ghaw.